q/qtool: better -kernel boots
The old cmdline for q is not that easy to use, make it as simple as
generating a common QEMU cmdline that mostly I'll use when invoking a VM,
then build base on that.

Add cmds for qtool to either boot a fresh kernel or commit image data to
backing store.

Signed-off-by: Peter Xu <peterx@redhat.com>